Modularize getting all Tags
