+++ apps/codeforafrica/src/lib/data/blockify/opportunities.js
@@ -1,4 +1,7 @@
-import { getPosts, formatTags } from "@/codeforafrica/lib/data/utils/posts";
+import {
+  getPosts,
+  getTagsByPrimaryTag,
+} from "@/codeforafrica/lib/data/utils/posts";
 
 async function opportunities(block, api, context) {
   const primaryTag = "opportunities";
@@ -11,21 +14,7 @@ async function opportunities(block, api, context) {
 
   const { posts, pagination } = await getPosts(api, options, primaryTag);
 
-  const { docs: allOpportunities } = await api.getCollection("posts", {
-    limit: 0,
-    where: {
-      "tags.name": {
-        like: primaryTag,
-      },
-    },
-  });
-
-  const allTags = allOpportunities.reduce((acc, story) => {
-    const { tags = [] } = story;
-    return [...acc, ...tags.map(({ name, slug }) => ({ name, slug }))];
-  }, []);
-
-  const tags = formatTags(allTags);
+  const tags = await getTagsByPrimaryTag(api, primaryTag);
 
   return {
     labels,
